Support updating classification when a document's project context changes.
