+++ src/EditorFeatures/Core/Implementation/Classification/SyntacticClassificationTaggerProvider.TagComputer.cs
@@ -2,6 +2,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Classification;
@@ -391,6 +392,9 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.Classification
             {
                 if (_workspace != null)
                 {
+                    // If the document's active context changes, then we can't cache anything.
+                    // All the data may now be invalid.
+                    _lastLineCache.Clear();
                     ParseIfThisDocument(null, args.Document.Project.Solution, args.Document.Id);
                 }
             }
@@ -503,10 +507,11 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.Classification
             {
                 if (_workspace != null)
                 {
-                    var openDocumentId = _workspace.GetDocumentIdInCurrentContext(_subjectBuffer.AsTextContainer());
-                    if (openDocumentId == documentId)
+                    var relatedDocuments = _workspace.GetRelatedDocumentIds(_subjectBuffer.AsTextContainer());
+                    if (relatedDocuments.Contains(documentId))
                     {
-                        EnqueueParseSnapshotTask(newSolution.GetDocument(documentId));
+                        var documentIdInCurrentContext = _workspace.GetDocumentIdInCurrentContext(documentId);
+                        EnqueueParseSnapshotTask(newSolution.GetDocument(documentIdInCurrentContext));
                     }
                 }
             }
